changelog

+++ deepinv/models/tv.py
@@ -145,7 +145,7 @@ class TVDenoiser(Denoiser):
         Applies the finite differences operator associated with tensors of the same shape as x.
         """
         b, c, h, w = x.shape
-        u = torch.zeros((b, c, h, w, 2), device=x.device).type(x.dtype)
+        u = torch.zeros((b, c, h, w, 2), device=x.device, dtype=x.dtype)
         u[:, :, :-1, :, 0] = u[:, :, :-1, :, 0] - x[:, :, :-1]
         u[:, :, :-1, :, 0] = u[:, :, :-1, :, 0] + x[:, :, 1:]
         u[:, :, :, :-1, 1] = u[:, :, :, :-1, 1] - x[..., :-1]
@@ -158,9 +158,8 @@ class TVDenoiser(Denoiser):
         Applies the adjoint of the finite difference operator.
         """
         b, c, h, w = x.shape[:-1]
-        u = torch.zeros((b, c, h, w), device=x.device).type(
-            x.dtype
-        )  # note that we just reversed left and right sides of each line to obtain the transposed operator
+        u = torch.zeros((b, c, h, w), device=x.device, dtype=x.dtype)
+        # note that we just reversed left and right sides of each line to obtain the transposed operator
         u[:, :, :-1] = u[:, :, :-1] - x[:, :, :-1, :, 0]
         u[:, :, 1:] = u[:, :, 1:] + x[:, :, :-1, :, 0]
         u[..., :-1] = u[..., :-1] - x[..., :-1, 1]
